first pass on moving views to individual apps
